+++ src/DemoUI/src/Views/ACME/index.php
@@ -225,7 +225,9 @@ echo $lang ?>">
                     startAutoConfigureUrl: "<?php getUrl('AutoConfigure', 'start') ?>",
                     disableCarriersUrl: "<?php getUrl('ShippingMethods', 'deactivate') ?>",
                     getCurrencyDetailsUrl: "<?php getUrl('SystemInfo', 'get') ?>",
-                    systemId: null
+                    systemId: null,
+                    enqueue: "<?php getUrl('ManualRefresh', 'enqueueUpdateTask'); ?>",
+                    getTaskStatus: "<?php getUrl('ManualRefresh', 'getTaskStatus'); ?>"
                 },
                 'edit-service': {
                     getServiceUrl: "<?php getUrl('ShippingMethods', 'getService') ?>",